cleaning up styling and adding quality of life
I am working on improving how things are positioned, specifically by playing with the display, margins, padding, and containers. I added a dialog window that forces you to confirm before deleting a transaction. I refit the text inside of the transactions, adding a divider between the name and amount. I added an info icon that contains a tooltip with the rest of the relevant transaction details. That still needs work to finish formatting the display of the information. I also improved the design of the transaction delete button.