corrected dialog focus and began adding inputs
The issue with the dialog was that I was passing a new instance of the transition to the dialog each time. That was breaking it?? so I moved the transition const outside of the function itself.

I also began to modify the dialog to contain input fields for the new transaction.