added backend code and reorganized file pathing
I took the java backend code from mindful and dropped it into this project. that code will need most of it ripped out but I am using it to test the fetch api and I will likely use the authentication code.

I broke the file organization down to frontend and java