extracted transaction deletion from transaction component
I removed the delete button and confirmation dialog out of the transaction component. With this, the isOpen state is no longer inside of the transaction so opening the dialog will no longer cause the entire transaction list and import button to re-render. This delete confirmation can potentially be used elsewhere in the application as well.