fixed transaction form bug
The form was not closing and the page was resetting bc I never finished the functions to import/create transactions lol I needed to add a prevent default to the form submit when importing and they both needed to call the isOpen function when submitting. That function coming from the parent (which the parent needs to be extracted out from the button speed dial into its own component) controls the state of the form being open.